requirements.txt
Flask==2.2.2
PyYAML==6.0
Werkzeug==2.2.2
mysql-connector-python~=8.0.31